Create Twitter Analysis - NLP
COVID-19 has generated a plethora of discussions on social media. Twitter has been extensively
used as a platform to discuss facts, share opinions, and express sentiments during the times of
crisis. It is imperative to understand that the response to COVID-19 in different areas may be
dependent upon local cases, government policies, health infrastructure and multiple
socioeconomic factors. As the pandemic has affected everyone’s’ life on a different scale, it is
essential to capture the public’s main topics of concern, and how people in different locations
have reacted to the situation. In the event of crisis, monitoring people's responses across various
events and topics will be extremely necessary for policy makers to be able to devise effective
policies. In this dissertation, twitter data is used to perform region-specific exploratory analysis
on people’s discussion and sentiments about COVID-19 in the UK. Topic modelling and
sentiment analysis is used to unravel the intra-region dynamics. It shows that people across
different regions discuss the same topics in different contexts. The study also discusses how
sentiments of people change over time and are heavily influenced by government policies and
information on media. Overall, the proposed framework shows the potential of using twitter
analysis to assist the local authorities in designing policies for new interventions.